Fixed the href and edit some texts.
Hanni page did not work when clicked.
The problem was the href value, which was New Jeans but I made it NewJeans.

Spelling from Danielle page has errors.
Made some bold, italicized, marked and underlined texts for a bit of fun.
It was not really consistent. It was to experiment.